+++ packages/changed-elements-react/src/widgets/comparisonJobWidget/models/NamedVersions.ts
@@ -2,7 +2,7 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import { VersionState } from "./VersionState";
+import type { VersionState } from "./VersionState.js";
 
 /**
  * Holds the version state of named versions and the current version.